+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/App/AdminPanel.Client.App.csproj
@@ -59,7 +59,7 @@
     <ItemGroup>
 
         <Using Include="Microsoft.AspNetCore.Components.WebView.Maui" />
-        <Using Include="AdminPanel.Client.Core.Components" />
+        <Using Include="AdminPanel.Client.Core.Shared" />
         <Using Include="AdminPanel.Client.Core.Services.Contracts" />
         <Using Include="AdminPanel.Client.Core.Services.Implementations" />
