+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/AdminPanel.Client.Core.csproj
@@ -45,7 +45,6 @@
         <Using Include="Bit.BlazorUI" />
         <Using Include="AdminPanel.Shared" />
         <Using Include="AdminPanel.Client.Core.Shared" />
-        <Using Include="AdminPanel.Client.Core.Components" />
         <Using Include="AdminPanel.Client.Core.Services.Contracts" />
         <Using Include="AdminPanel.Client.Core.Services.Implementations" />
