+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Categories/AddOrEditCategoryPage.razor
@@ -1,5 +1,5 @@
 ﻿@page "/add-edit-category/{Id:int?}"
-@namespace AdminPanel.Client.Core.Pages
+
 @inherits AppComponentBase
 
 <PageTitle>
