+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Categories/AddOrEditCategoryPage.razor.cs
@@ -1,6 +1,6 @@
 ﻿using AdminPanel.Shared.Dtos.Categories;
 
-namespace AdminPanel.Client.Core.Pages;
+namespace AdminPanel.Client.Core.Pages.Categories;
 
 [Authorize]
 public partial class AddOrEditCategoryPage
