+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Categories/CategoriesPage.razor
@@ -1,6 +1,5 @@
 ﻿@page "/categories"
 @using AdminPanel.Shared.Dtos.Categories
-@namespace AdminPanel.Client.Core.Pages
 @inherits AppComponentBase
 
 <PageTitle>@Localizer[nameof(AppStrings.CategoriesPageTitle)]</PageTitle>
