+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Home/HomePage.razor
@@ -1,5 +1,4 @@
 ﻿@page "/"
-@namespace AdminPanel.Client.Core.Pages
 @inherits AppComponentBase
 
 <div class="page-container">
