+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Home/HomePage.razor.cs
@@ -1,4 +1,4 @@
-﻿namespace AdminPanel.Client.Core.Pages;
+﻿namespace AdminPanel.Client.Core.Pages.Home;
 
 [Authorize]
 public partial class HomePage
