+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Home/ProductsPercentageWidget.razor.cs
@@ -1,4 +1,4 @@
-﻿namespace AdminPanel.Client.Core.Pages;
+﻿namespace AdminPanel.Client.Core.Pages.Home;
 
 public partial class ProductsPercentageWidget
 {
