+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Home/ProductsSalesWidget.razor
@@ -1,5 +1,4 @@
-﻿@namespace AdminPanel.Client.Core.Pages
-@inherits AppComponentBase
+﻿@inherits AppComponentBase
 
 <div class="card">
     <div class="card-header">
