+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Identity/DeleteAccountConfirmModal.razor
@@ -1,5 +1,4 @@
-﻿@namespace AdminPanel.Client.Core
-@inherits AppComponentBase
+﻿@inherits AppComponentBase
 
 <BitModal @bind-IsOpen="IsOpen">
     <div class="main-container">
