+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Identity/DeleteAccountConfirmModal.razor.scss
