+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Identity/EditProfilePage.razor
@@ -1,5 +1,4 @@
 ﻿@page "/edit-profile"
-@namespace AdminPanel.Client.Core.Pages
 @inherits AppComponentBase
 
 <PageTitle>@Localizer[nameof(AppStrings.EditProfileTitle)]</PageTitle>
