+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Identity/EmailConfirmationPage.razor.scss
@@ -1,5 +1,6 @@
-﻿@import '../Styles/abstracts/_functions.scss';
-@import '../Styles/abstracts/_media-queries.scss';
+﻿@import '../../Styles/abstracts/_functions.scss';
+@import '../../Styles/abstracts/_media-queries.scss';
+@import '../../Styles/abstracts/_bit-css-variables.scss';
 
 .page-container {
     width: 100%;
@@ -28,16 +29,16 @@
 
     &:first-child {
         display: flex;
-        align-items: flex-start;
         justify-content: center;
+        align-items: flex-start;
         flex-flow: column nowrap;
         padding: rem2(135px) rem2(76px) rem2(40px);
 
         @include md {
             align-items: flex-end;
+            padding: rem2(25px) rem2(32px);
             flex-flow: row-reverse nowrap;
             justify-content: space-between;
-            padding: rem2(25px) rem2(32px);
         }
 
         @include sm {
@@ -118,6 +119,7 @@
 ::deep .form {
     width: 100%;
     display: flex;
+    text-align: center;
     align-items: center;
     max-width: rem2(384px);
     flex-flow: column nowrap;
@@ -152,7 +154,21 @@
     position: absolute;
 }
 
-::deep .reset-password-btn {
-    width: rem2(130px);
-    max-width: rem2(130px);
+.form-description {
+    font-weight: 600;
+    width: rem2(398px);
+    height: rem2(20px);
+    font-size: rem2(14px);
+    line-height: rem2(20px);
+    margin-bottom: rem2(25px);
+
+    @include md {
+        width: rem2(398px);
+        height: rem2(20px);
+    }
+
+    @include sm {
+        width: rem2(280px);
+        height: rem2(40px);
+    }
 }
