+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Identity/ForgotPasswordPage.razor.scss
@@ -1,13 +1,14 @@
-﻿@import '../Styles/abstracts/_functions.scss';
-@import '../Styles/abstracts/_media-queries.scss';
+﻿@import '../../Styles/abstracts/_functions.scss';
+@import '../../Styles/abstracts/_media-queries.scss';
+@import '../../Styles/abstracts/_bit-css-variables.scss';
 
 .page-container {
     width: 100%;
     height: 100%;
     display: flex;
     flex-flow: row nowrap;
-    align-items: flex-start;
     justify-content: center;
+    align-items: flex-start;
 
     @include md {
         align-items: center;
@@ -28,8 +29,8 @@
 
     &:first-child {
         display: flex;
-        justify-content: center;
         align-items: flex-start;
+        justify-content: center;
         flex-flow: column nowrap;
         padding: rem2(135px) rem2(76px) rem2(40px);
 
@@ -118,9 +119,8 @@
 ::deep .form {
     width: 100%;
     display: flex;
-    text-align: center;
     align-items: center;
-    max-width: rem2(405px);
+    max-width: rem2(384px);
     flex-flow: column nowrap;
     justify-content: flex-start;
 }
@@ -145,19 +145,22 @@
 
 .form-input-container {
     width: 100%;
-    text-align: left;
     margin-bottom: rem2(24px);
 }
 
-.sign-in-row {
-    font-weight: 400;
-    font-size: rem2(14px);
-    margin-top: rem2(24px);
-    line-height: rem2(20px);
+::deep .bit-msg-bar-fluent {
+    top: 0;
+    position: absolute;
 }
 
 .form-description {
-    font-weight: 600;
+    text-align: center;
     font-size: rem2(14px);
     line-height: rem2(20px);
+    margin-bottom: rem2(16px);
+}
+
+::deep .send-reset-link-btn {
+    width: rem2(130px);
+    max-width: rem2(130px);
 }
