+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Identity/ResetPasswordPage.razor.scss
@@ -1,5 +1,5 @@
-﻿@import '../Styles/abstracts/_functions.scss';
-@import '../Styles/abstracts/_media-queries.scss';
+﻿@import '../../Styles/abstracts/_functions.scss';
+@import '../../Styles/abstracts/_media-queries.scss';
 
 .page-container {
     width: 100%;
@@ -28,23 +28,19 @@
 
     &:first-child {
         display: flex;
-        justify-content: center;
         align-items: flex-start;
+        justify-content: center;
         flex-flow: column nowrap;
         padding: rem2(135px) rem2(76px) rem2(40px);
 
         @include md {
             align-items: flex-end;
-            padding: rem2(25px) rem2(32px);
             flex-flow: row-reverse nowrap;
             justify-content: space-between;
+            padding: rem2(25px) rem2(32px);
         }
 
         @include sm {
-            display: none;
-        }
-
-        @media screen and (max-height:700px) {
             display: none;
         }
     }
@@ -151,32 +147,12 @@
     margin-bottom: rem2(24px);
 }
 
-.sign-up-row {
-    font-weight: 400;
-    font-size: rem2(14px);
-    margin-top: rem2(24px);
-    line-height: rem2(20px);
-}
-
-.form-forgot-password {
-    font-weight: 400;
-    font-size: rem2(14px);
-    margin-top: rem2(24px);
-    line-height: rem2(20px);
-}
-
 ::deep .bit-msg-bar-fluent {
-    top: 24px;
+    top: 0;
     position: absolute;
-    width: calc(100% - 48px);
+}
 
-    @include md {
-        top: 16px;
-        width: calc(100% - 32px);
-    }
-
-    @include sm {
-        top: 16px;
-        width: calc(100% - 32px);
-    }
+::deep .reset-password-btn {
+    width: rem2(130px);
+    max-width: rem2(130px);
 }
