+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Products/CreateEditProductModal.razor
@@ -1,5 +1,4 @@
-﻿@namespace AdminPanel.Client.Core.Pages
-@inherits AppComponentBase
+﻿@inherits AppComponentBase
 
 <BitModal @bind-IsOpen="_isOpen">
     <div class="main-container">
