+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Products/ProductsPage.razor
@@ -1,6 +1,5 @@
 ﻿@page "/products"
 @using AdminPanel.Shared.Dtos.Products
-@namespace AdminPanel.Client.Core.Pages
 @inherits AppComponentBase
 
 <PageTitle>@Localizer[nameof(AppStrings.ProductsPageTitle)]</PageTitle>
