+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Pages/Products/ProductsPage.razor.cs
@@ -1,7 +1,7 @@
 ﻿//-:cnd:noEmit
 using AdminPanel.Shared.Dtos.Products;
 
-namespace AdminPanel.Client.Core.Pages;
+namespace AdminPanel.Client.Core.Pages.Products;
 
 [Authorize]
 public partial class ProductsPage
