+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Shared/AppComponentBase.cs
@@ -1,4 +1,4 @@
-﻿namespace AdminPanel.Client.Core.Components;
+﻿namespace AdminPanel.Client.Core.Shared;
 
 public partial class AppComponentBase : ComponentBase
 {
