+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/Shared/ConfirmMessageBox.razor.scss
