+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/_Imports.razor
@@ -16,6 +16,5 @@
 @using AdminPanel.Client.Core
 @using AdminPanel.Client.Core.Shared
 @using AdminPanel.Client.Core.Pages
-@using AdminPanel.Client.Core.Components
 @using AdminPanel.Client.Core.Services.Contracts
 @using AdminPanel.Client.Core.Services.Implementations
