+++ src/Templates/AdminPanel/Bit.AdminPanel/src/Client/Core/compilerconfig.json
@@ -1,7 +1,7 @@
 ﻿[
     {
-        "outputFile": "Components/ConfirmMessageBox.razor.css",
-        "inputFile": "Components/ConfirmMessageBox.razor.scss",
+        "outputFile": "Shared/ConfirmMessageBox.razor.css",
+        "inputFile": "Shared/ConfirmMessageBox.razor.scss",
         "minify": { "enabled": false },
         "options": { "sourceMap": false }
     },
@@ -126,32 +126,32 @@
         "options": { "sourceMap": false }
     },
     {
-        "outputFile": "Pages/EmailConfirmationPage.razor.css",
-        "inputFile": "Pages/EmailConfirmationPage.razor.scss",
+        "outputFile": "Pages/Identity/EmailConfirmationPage.razor.css",
+        "inputFile": "Pages/Identity/EmailConfirmationPage.razor.scss",
         "minify": { "enabled": false },
         "options": { "sourceMap": false }
     },
     {
-        "outputFile": "Pages/ForgotPasswordPage.razor.css",
-        "inputFile": "Pages/ForgotPasswordPage.razor.scss",
+        "outputFile": "Pages/Identity/ForgotPasswordPage.razor.css",
+        "inputFile": "Pages/Identity/ForgotPasswordPage.razor.scss",
         "minify": { "enabled": false },
         "options": { "sourceMap": false }
     },
     {
-        "outputFile": "Pages/ResetPasswordPage.razor.css",
-        "inputFile": "Pages/ResetPasswordPage.razor.scss",
+        "outputFile": "Pages/Identity/ResetPasswordPage.razor.css",
+        "inputFile": "Pages/Identity/ResetPasswordPage.razor.scss",
         "minify": { "enabled": false },
         "options": { "sourceMap": false }
     },
     {
-        "outputFile": "Pages/SignInPage.razor.css",
-        "inputFile": "Pages/SignInPage.razor.scss",
+        "outputFile": "Pages/Identity/SignInPage.razor.css",
+        "inputFile": "Pages/Identity/SignInPage.razor.scss",
         "minify": { "enabled": false },
         "options": { "sourceMap": false }
     },
     {
-        "outputFile": "Pages/SignUpPage.razor.css",
-        "inputFile": "Pages/SignUpPage.razor.scss",
+        "outputFile": "Pages/Identity/SignUpPage.razor.css",
+        "inputFile": "Pages/Identity/SignUpPage.razor.scss",
         "minify": { "enabled": false },
         "options": { "sourceMap": false }
     },
